requirements.txt
streamlit
openai
python-docx
python-dotenv
requests